Edited the Swedish layer
Entering é takes three keypresses, which has now been changed to
two. å and Å has been moved to the left hand, since I appear
to favor the right when triggering the OSM layer, and it makes
sense for å to live under the a, while ö continues to live under
o and ä under e. Finally, changed the enum-symbol for the acute
key.